refactor: move feature toggle store
