+++ src/lib/db/feature-strategy-store.test.ts
@@ -1,7 +1,7 @@
 import dbInit from '../../test/e2e/helpers/database-init';
 import getLogger from '../../test/fixtures/no-logger';
 import FeatureStrategiesStore from './feature-strategy-store';
-import FeatureToggleStore from './feature-toggle-store';
+import FeatureToggleStore from '../features/feature-toggle/feature-toggle-store';
 import StrategyStore from './strategy-store';
 import { IFeatureStrategy, PartialSome } from '../types';
