+++ src/lib/features/project/createProjectService.ts
@@ -13,7 +13,7 @@ import {
 import FakeGroupStore from '../../../test/fixtures/fake-group-store';
 import FakeEventStore from '../../../test/fixtures/fake-event-store';
 import ProjectStore from '../../db/project-store';
-import FeatureToggleStore from '../../db/feature-toggle-store';
+import FeatureToggleStore from '../feature-toggle/feature-toggle-store';
 import FeatureTypeStore from '../../db/feature-type-store';
 import { FeatureEnvironmentStore } from '../../db/feature-environment-store';
 import ProjectStatsStore from '../../db/project-stats-store';
