+++ src/lib/services/client-metrics/last-seen/createLastSeenService.ts
@@ -1,5 +1,5 @@
 import FakeFeatureToggleStore from '../../../../test/fixtures/fake-feature-toggle-store';
-import FeatureToggleStore from '../../../db/feature-toggle-store';
+import FeatureToggleStore from '../../../features/feature-toggle/feature-toggle-store';
 import { Db, IUnleashConfig } from '../../../server-impl';
 import { FakeLastSeenStore } from './fake-last-seen-store';
 import { LastSeenService } from './last-seen-service';
