+++ src/test/fixtures/fake-feature-toggle-store.ts
@@ -10,7 +10,7 @@ import {
     IVariant,
 } from 'lib/types/model';
 import { LastSeenInput } from '../../lib/services/client-metrics/last-seen/last-seen-service';
-import { EnvironmentFeatureNames } from '../../lib/db/feature-toggle-store';
+import { EnvironmentFeatureNames } from '../../lib/features/feature-toggle/feature-toggle-store';
 
 export default class FakeFeatureToggleStore implements IFeatureToggleStore {
     features: FeatureToggle[] = [];
